- Adding Committee model to admin site register - Adding ManyToMany relation to committees to a bill. That way if a user wants to see exactly what a committee is up to they can see the bills they're involved with. - Adding more fields to the Committee model to improve relativity and lookups. I also added nullability to the type and chamber fields because they're not immediately present when I process an action. Thus when I actually find the Committee's full data later I can add the other fields to the existing model (the action lists the system code and name of the committee) to add it to the bill. - Abstracting legislator lookups into a helper method called get_legislator. It doesn't necessarily add any performance benefits, but it was the same code twice for looking up a legislator for cosponsorships and sponsorships, so I made it into a separate method to reduce redundancy. - Added in the parsing of committees from the XML document and updating or creating them based on the context. If I haven't seen the committee's full data yet, then I put the full data into the object.
At this point in time I think that I'm fully parsing the data I need to display stuff to a user in the iOS application. My next big transition will be creating API endpoints so a user can access the data. A big todo in the future for the data I'm currently storing will be looking up the listed CBO document links as well as senate and house roll calls. This will involve a whole other method of parsing that I haven't encountered yet, so I'm going to put it on the back burner while I develop the API endpoints.